backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# Database
motor>=3.5.0
pymongo>=4.8.0
beanie>=1.26.0

# Google Gemini AI
google-generativeai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Validation (with Python 3.14 support)
pydantic>=2.12.0
pydantic-settings>=2.6.0

# Date/Time handling
python-dateutil>=2.9.0

# Optional: Testing
pytest>=8.0.0
httpx>=0.27.0
